docker image on mvn pkg cmd

+++ customer/src/main/java/com/code365/customer/repository/CustomerRepository.java
@@ -1,8 +1,11 @@
 package com.code365.customer.repository;
 
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.code365.customer.model.Customer;
+
 @Repository
-public interface CustomerRepository {//extends JpaRepository<Customer, Integer>{
+public interface CustomerRepository {// extends JpaRepository<Customer, Long> {
 
 }
